Go, Git, Bash, network, Postgresql
Go: a separate readme is created.

Git: designation of RSA key.

Bash: Run an SSH command stored in a variable. Compute amount of space dedicated for file of a certain pattern. Add user and add him to sudo.

Network: tcpdump shows data of a package and saves output to a file. Setup logging in iptables.

Postgresql: notes on privelleges.